pi-app/requirements.txt
fastapi
uvicorn
python-multipart
paho-mqtt